add GUI - boards part done with a very good resizing effect. add FXML files, image resources, fx controller and a application tester. (NB. resources may require a bit more time to update)

+++ src/main/java/porprezhas/control/GameController.java
@@ -210,7 +210,7 @@ public class GameController implements GameControllerInterface, Runnable {
         while(!bChosen) {
             try {
                 synchronized(chooseTimeOut) {
-                    chooseTimeOut.wait( (int) Game.GameConstants.secondsToMillis(
+                    chooseTimeOut.wait( Game.GameConstants.secondsToMillis(
                             Game.GameConstants.TIMEOUT_PREPARING_SEC));
                     chooseTimeOut.notifyAll();
                     setup();
