+++ src/main/java/porprezhas/model/GameInterface.java
@@ -3,7 +3,6 @@ package porprezhas.model;
 import porprezhas.model.dices.DiceBag;
 import porprezhas.model.dices.DraftPool;
 
-import java.rmi.RemoteException;
 import java.util.List;
 
 // To create a new Game needs a PlayerList as parameter for construction method
@@ -19,7 +18,7 @@ public interface GameInterface {
 
     boolean InsertDice(int indexDice, int xPose, int yPose);
     boolean isSolitaire();
-    int getRoundTimeOut();
+    long getRoundTimeOut();
     boolean setPattern(Player player, int indexPatternType);
     void nextRound();
     int calcScore(Player player);
